+++ CHANGES.md
@@ -1,6 +1,7 @@
 ## In the next release...
 
 * Add experimental StatsD telemeter
+* Add a recent requests log in the linkerd admin dashboard
 
 ## 0.8.5 2017-01-06